+++ src/js/validator/digits.js
@@ -9,7 +9,12 @@
          * @returns {Boolean}
          */
         validate: function(validator, $field, options) {
-            return /^\d+$/.test($field.val());
+            var value = $field.val();
+            if (value == '') {
+                return true;
+            }
+
+            return /^\d+$/.test(value);
         }
     }
 }(window.jQuery));