+++ src/js/validator/lessThan.js
@@ -12,7 +12,11 @@
          * @returns {Boolean}
          */
         validate: function(validator, $field, options) {
-            var value = parseFloat($field.val());
+            var value = $field.val();
+            if (value == '') {
+                return true;
+            }
+            value = parseFloat(value);
             return (options.inclusive === true) ? (value < options.value) : (value <= options.value);
         }
     };