+++ src/js/validator/stringLength.js
@@ -12,7 +12,12 @@
          * @returns {Boolean}
          */
         validate: function(validator, $field, options) {
-            var value = $.trim($field.val()), length = value.length;
+            var value = $field.val();
+            if (value == '') {
+                return true;
+            }
+
+            var length = $.trim(value).length;
             if ((options.min && length < options.min) || (options.max && length > options.max)) {
                 return false;
             }